feat(matcher): E1.3b — move configs + convertors + receptor_hierarchy into exact_matching
Second slice of E1.3. Moves the matcher's remaining pure-logic subtree (no
trials.models / ORM refs) into the package + shims:
- patient_info/configs.py (USER_TO_TRIAL_ATTRS_MAPPING & friends)
- patient_info/convertors/ (12 modules — ULN calculators, unit convertors)
- receptor_hierarchy.py

Internal edges repoint to package paths so exact_matching is self-contained on
this subtree: configs imports its convertors + receptor_hierarchy from the
package; convertors import each other from the package. The matcher's configs
import repoints too — leaving patient_info_attributes as its ONLY remaining
trials dep (E1.3c). django-pint declared as a package dep (base_convertor uses
quantityfield).

Old paths are import-* shims (one per convertor module — models.py, normalize,
patient_info_attributes import them by full path). External importers unchanged:
configs ×8, convertors ×3, receptor_hierarchy ×2.

Full EXACT suite green (1073 passed, 5 skipped).

Part of EPIC cancerbot#4601 (E1 #272, D1 resolved). Proposal — cross-team.
