+++ matcher_app/pyproject.toml
@@ -10,7 +10,7 @@ name = "exact-matching"
 version = "0.0.1"
 description = "Shared CB/EXACT trial matcher (queryset + per-trial matcher + config), behind CB's MatcherBackend interface. Extraction in progress (E1)."
 requires-python = ">=3.12"
-dependencies = ["Django>=5.0", "glom>=24.0"]
+dependencies = ["Django>=5.0", "glom>=24.0", "django-pint>=1.0"]
 
 [tool.setuptools.packages.find]
 include = ["exact_matching*"]
