+++ trials/services/patient_info/convertors/__init__.py
@@ -1,0 +1,1 @@
+"""Back-compat shim package — modules moved to exact_matching.patient_info.convertors (E1.3b)."""
